refactor(data): move VLM collators to model modules
