+++ src/megatron/bridge/models/nemotron_vl/data/collate_fn.py
@@ -1,0 +1,123 @@
+# Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Nemotron VL collator implementations."""
+
+import torch
+
+from megatron.bridge.data.datasets.utils import IGNORE_INDEX
+from megatron.bridge.data.vlm_datasets.token_utils import extract_skipped_token_ids
+from megatron.bridge.data.vlm_processing import build_assistant_loss_mask
+from megatron.bridge.training.utils.visual_inputs import GenericVisualInputs
+
+
+def nemotron_nano_v2_vl_collate_fn(examples: list, processor, start_of_response_token=None) -> dict[str, torch.Tensor]:
+    """Collate function for Nemotron Nano V2 VL model."""
+    from megatron.bridge.models.nemotron_vl.nemotron_vl_utils import adjust_image_tokens
+
+    skipped_tokens = extract_skipped_token_ids(processor)
+    # this assumes the first message in conversation is the video message
+    is_video = examples[0]["conversation"][0]["content"][0]["type"] == "video"
+    if is_video:
+        from megatron.bridge.models.nemotron_vl.nemotron_vl_utils import (
+            maybe_path_or_url_to_data_urls,
+            pil_image_from_base64,
+        )
+
+        assert len(examples) == 1, "Nemotron Nano V2 VL processor only supports batch size == 1"
+        frames = []
+        video_fps = -1
+        video_nframe = 10
+        video_nframe_max = -1
+
+        for example in examples:
+            video_path = example["conversation"][0]["content"][0]["path"]
+            image_urls, metadata = maybe_path_or_url_to_data_urls(
+                video_path,
+                fps=max(0, int(video_fps)),
+                nframe=max(0, int(video_nframe)),
+                nframe_max=int(video_nframe_max),
+            )
+            frames.append([pil_image_from_base64(image_url) for image_url in image_urls])
+
+        prompt = processor.apply_chat_template([example["conversation"] for example in examples], tokenize=False)
+        batch = processor(
+            text=prompt,
+            videos=frames,
+            videos_kwargs={"video_metadata": metadata},
+            return_tensors="pt",
+        )
+    else:
+        # Ensure a pad_token is set so padding can produce uniform-length tensors.
+        if processor.tokenizer.pad_token is None:
+            processor.tokenizer.pad_token = processor.tokenizer.eos_token
+        batch = processor.apply_chat_template(
+            [example["conversation"] for example in examples],
+            tokenize=True,
+            padding=True,
+            truncation=True,
+            return_tensors="pt",
+            return_dict=True,
+        )
+    loss_mask = torch.stack(
+        [
+            build_assistant_loss_mask(example, input_ids, processor, skipped_tokens).to(dtype=torch.int)
+            for example, input_ids in zip(examples, batch["input_ids"])  # type: ignore[arg-type]
+        ]
+    )
+
+    img_start_token_id = 131073  # tokenizer.convert_tokens_to_ids("<img>")
+    img_end_token_id = 131074  # tokenizer.convert_tokens_to_ids("</img>")
+    adjusted_batch = adjust_image_tokens(
+        {
+            "input_ids": batch["input_ids"],
+            "loss_mask": loss_mask,
+        },
+        batch["num_patches"],
+        img_start_token_id,
+        img_end_token_id,
+    )
+
+    if is_video:
+        video_token_id = processor.tokenizer.convert_tokens_to_ids("<video>")
+        image_token_id = processor.tokenizer.convert_tokens_to_ids("<image>")
+        adjusted_batch["input_ids"] = torch.where(
+            adjusted_batch["input_ids"] == video_token_id, image_token_id, adjusted_batch["input_ids"]
+        )
+
+    batch["input_ids"] = adjusted_batch["input_ids"]
+    loss_mask = adjusted_batch["loss_mask"]
+
+    if "position_ids" not in batch:
+        batch_size, seq_len = batch["input_ids"].shape
+        batch["position_ids"] = (
+            torch.arange(seq_len, device=batch["input_ids"].device).unsqueeze(0).expand(batch_size, -1)
+        )
+
+    key = "pixel_values_videos" if is_video else "pixel_values"
+    pv = batch[key].to(torch.bfloat16)
+    batch[key] = pv
+    batch["visual_inputs"] = GenericVisualInputs(pixel_values=pv)
+    # roll label by 1 and fill last token with IGNORE_INDEX
+    labels = batch["input_ids"].clone()[:, 1:]
+    labels = torch.cat([labels, IGNORE_INDEX * torch.ones_like(labels[:, :1])], dim=1)
+    labels[torch.isin(labels, skipped_tokens)] = IGNORE_INDEX
+    batch["labels"] = labels
+
+    loss_mask_t = loss_mask.to(dtype=torch.float, device=batch["input_ids"].device)
+    # Shift loss mask to align with next-token labels timeline
+    loss_mask_t = torch.cat([loss_mask_t[:, 1:], torch.zeros_like(loss_mask_t[:, :1])], dim=1)
+    batch["labels"] = batch["labels"].masked_fill(loss_mask_t == 0, IGNORE_INDEX)
+    batch["loss_mask"] = loss_mask_t
+    return batch
